contrib/pam_ssh_agent_auth: Add PAM module for ssh-agent authentication.
This module can be used to gain sudo access via SSH keys.  The module
checks the available keys from a connected SSH agent against a list of
authorized keys in /etc/security/authorized_keys.
If any of the keys is authorized the PAM module generates a challenge
consisting of a fixed string and a 32 byte random nonce to be signed by
the connected SSH agent. If the signature passes verification,
access is granted.

Based on the original implementation by Jamie Beverly.